hccr/requirements.txt
numpy==1.12.1
Pillow==2.3.0
requests==2.2.1
Django==1.11.3
gunicorn==19.7.1
whitenoise==3.3.0
https://storage.googleapis.com/tensorflow/linux/cpu/tensorflow-1.2.1-cp36-cp36m-linux_x86_64.whl